Prepare for Rust Migration
