--- .vscode/c_cpp_properties.json
@@ -1,17 +1,0 @@
-{
-    "configurations": [
-        {
-            "name": "Linux",
-            "includePath": [
-                "${default}"
-            ],
-            "defines": [],
-            "compilerPath": "/usr/bin/gcc",
-            "intelliSenseMode": "linux-gcc-x64",
-            "configurationProvider": "ms-vscode.cmake-tools",
-            "cStandard": "c11",
-            "cppStandard": "c++14"
-        }
-    ],
-    "version": 4
-}
