--- .vscode/settings.json
@@ -1,19 +1,0 @@
-{
-	"files.associations": {
-		"*.meta": "ini",
-		"DIALOG.C": "cpp",
-		"TEXTURE.C": "cpp",
-		"SSIMAGE.C": "cpp",
-		"MATERIAL.C": "cpp",
-		"MATH.C": "cpp",
-		"SSA8.C": "cpp",
-		"COLOR.C": "cpp",
-		"FASTDIB.C": "cpp",
-		"SSDIB.C": "cpp",
-		"constants.h": "c"
-	},
-	"doxdocgen.file.copyrightTag": [
-		"@copyright Copyright (c) {year}. Work is based on original work from Microsoft Corp (c) 1994"
-	],
-	"C_Cpp.default.configurationProvider": "ms-vscode.cmake-tools"
-}
