+++ src-old/include/KHR/khrplatform.h
