+++ src-old/scratch/glTesting.cpp
